requirements.txt
solana
python-dotenv
numpy
pandas
requests